% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{base_linear_regression}
\alias{base_linear_regression}
\title{Title}
\usage{
base_linear_regression(
  Data,
  target,
  model_type,
  id_name = NULL,
  use_features = NULL
)
}
\arguments{
\item{Data}{Data to use the model}

\item{target}{dependent variable.}

\item{model_type}{which model to use: "lm" or "logistic" or "poisson"}

\item{id_name}{if Data has ID columns, select it...}

\item{use_features}{a character vector to use as features}
}
\value{
model
}
\description{
Title
}
