% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{changeStatus}
\alias{changeStatus}
\title{Change node status}
\usage{
changeStatus(id = NULL, status = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{node reference id}

\item{status}{one of none, queued, running, completed or error}

\item{session}{Shiny session}
}
\description{
Change the status of a node.  Status is shown by changing the border around the node.
By default the node status can be one of:
\itemize{
  \item none - no additional highlighting
  \item queued - yellow highlight
  \item running - green pulsing highlight
  \item completed - green highlight
  \item error - red pulsing highlight
}
}

