% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_pipeline_builder.R
\name{check_predict_method_throw_error}
\alias{check_predict_method_throw_error}
\title{Validate ml_pipeline_builder transform method returns data.frame}
\usage{
check_predict_method_throw_error(func_return_object)
}
\arguments{
\item{func_return_object}{The object returned from a \code{ml_pipeline_builder} method.}

\item{func_name}{The name of the function that returned the object.}
}
\description{
Helper function that checks if the object returned from a \code{ml_pipeline_builder} method is
data.frame (if it isn't NULL), and if it isn't, throws an error that is customised with the
returning name.
}
\examples{
\dontrun{
estimation_method <- function(df) lm(eruptions ~ 0 + waiting, df)
data <- faithful
model_estimate <- estimation_method(data)
check_predict_method_throw_error(model_estimate)
# NULL
}
}

