% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{checkfnList}
\alias{checkfnList}
\title{Check that function is in the function list}
\usage{
checkfnList(fn)
}
\arguments{
\item{fn}{Name of the function to run.}
}
\description{
Ensure that the node is included in the evaluation list
}

