% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to PostgreSQL}
\usage{
connect_db()
}
\value{
A DBI connection object (S4 class \code{"PqConnection"}).
}
\description{
Opens a connection with \code{\link[RPostgres:Postgres]{RPostgres::Postgres()}} using credentials from the
environment. All of \code{PG_DB}, \code{PG_HOST}, \code{PG_USER}, \code{PG_PASSWORD}, and
\code{PG_SCHEMA} must be set and non-empty; otherwise an error is raised before
connecting. Port is fixed at \code{5432}.
}
\section{Environment variables}{

\describe{
\item{PG_DB}{Database name.}
\item{PG_HOST}{Host name or IP.}
\item{PG_USER}{User name.}
\item{PG_PASSWORD}{Password.}
\item{PG_SCHEMA}{Schema used by \code{\link[=insert_new_data]{insert_new_data()}} and \code{\link[=push_summary_table]{push_summary_table()}}.}
}
}

\examples{
\dontrun{
con <- connect_db()
DBI::dbListTables(con)
DBI::dbDisconnect(con)
}
}
\seealso{
\code{\link[=start_pipeline]{start_pipeline()}}, \code{\link[=fetch_symbols]{fetch_symbols()}}, \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}
}
