% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_proj}
\alias{create_proj}
\title{Create project/processer/model (/experiment)}
\usage{
create_proj(project_name)
}
\arguments{
\item{project_name}{Target project name.}

\item{processer_name}{Target processer name.}

\item{model_name}{Target model name.}
}
\description{
Create project & manage experiment.
}
