% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{createNodes}
\alias{createNodes}
\title{Add a series of nodes to stencil canvas}
\usage{
createNodes(x = 0, y = 0, yOffset = 30, name = list(),
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{x}{position of first node in canvas (x-coordinate) defined in px}

\item{y}{position of first node in canvas (y-coordinate) defined in px}

\item{name}{list of node names (displayed as labels on the nodes)}

\item{session}{Shiny session}

\item{offset}{node offset in y-direction (in px).  Each node will be placed
\code{offset px} below the previous one}

\item{ports}{vector defining number of inputs (in) and outputs (out) for the node}
}
\description{
Add multiple nodes to the stencil canvas
}

