% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{deleteButton}
\alias{deleteButton}
\title{Change state of delete button}
\usage{
deleteButton(id = NULL, state = TRUE,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{node reference id}

\item{state}{true = visible, false = hidden}

\item{session}{Shiny session}
}
\description{
Set delete button to visible or hidden
}

