% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{executeNode}
\alias{executeNode}
\title{Execute a node}
\usage{
executeNode(fn = NULL, params = list())
}
\arguments{
\item{fn}{Name of the function to run.}

\item{params}{List of parameters to pass to the function.}
}
\value{
List containing output from the function.  First element (\code{result})
will be \code{error} or \code{success} and second element (\code{output}) is
an error report or the output of the node execution.
}
\description{
This is a function used to execute a node in the Cobralingus engine.
The code outputs a list containing two elements.  The first reports on the
success of running the node (\code{error} or \code{success}) and the second
contains either an error report or the output of the node execution.
}
\examples{
\dontrun{
executeNode('scrambleNouns', list(text, 50))
}

}

