% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSymbols.R
\name{fetchSymbols}
\alias{fetchSymbols}
\title{Extract a data.table object containing the symbols from to the ADEM PostgreSQL Database}
\usage{
fetchSymbols()
}
\value{
A data.table object
}
\description{
Extract the column 'symbols' from the adem db
}
\details{
Environment variables expected:
\itemize{
\item A data.table object named 'data'
}
}
\examples{
\dontrun{
symbolsDb <- fetchSymbols()
}
}
