% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{highlight}
\alias{highlight}
\title{Highlight a node}
\usage{
highlight(jnt = NULL, id = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{jnt}{reference to htmlwidget}

\item{id}{node reference id}

\item{session}{Shiny session}
}
\description{
Highlight a specific node with a box
}

