% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{includePackages}
\alias{includePackages}
\title{Add package names to node inclusion list}
\usage{
includePackages(pkgs = NULL)
}
\description{
Add a series of package names to the node inclusion list.
This defines the functions that can be executed within the workflow.
The function needs to be run before any nodes execute.  It creates two pkg.env
variables: \code{pkg.env$packageFunctions} which contains a list of nodes by package
and \code{pkg.env$allNodes} which contains all the allowed nodes
}

