% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_read}
\alias{input_read}
\title{Controll input data : read}
\usage{
input_read(processer_name)
}
\arguments{
\item{processer_name}{Target processer name.}
}
\description{
Controll input data. Read input data by processer name: file type is only csv.
}
