% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{isDAG}
\alias{isDAG}
\title{Check if pipeline is a directed acyclic graph}
\usage{
isDAG(jnt = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{jnt}{reference to htmlwidget}

\item{session}{Shiny session}
}
\description{
Determine if pipeline is a directed acyclic graph.  If graph is cyclic there's a danger of
getting caught in an infinite loop.
}

