% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{jntModule}
\alias{jntModule}
\title{shiny module server code to create a pipeline}
\usage{
jntModule(input, output, session, l.nodeTypes)
}
\arguments{
\item{l.nodeTypes}{List of nodes to include (comes from nodeParameters function)}
}
\description{
shiny module server code to create a pipeline
}

