% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointCanvas-v1.R
\name{jointCanvas}
\alias{jointCanvas}
\title{<Add Title>}
\usage{
jointCanvas(gridSize = 10, markAvailable = TRUE,
  restrictTranslate = FALSE, multiLinks = FALSE, border = FALSE,
  width = NULL, height = NULL)
}
\description{
<Add Description>
}

