% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processer.R
\docType{data}
\name{meta_dir}
\alias{meta_dir}
\title{Create processer}
\format{
An object of class \code{character} of length 1.
}
\usage{
meta_dir
}
\arguments{
\item{processer_name}{Target processer name.}
}
\description{
Create processer (Add processer script & model info to proc.yaml)
}
\keyword{datasets}
