% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{meta_mod_add}
\alias{meta_mod_add}
\title{Add meta data of model}
\usage{
meta_mod_add(model_name, processer_name)
}
\arguments{
\item{model_name}{Target model name.}

\item{processer_name}{Target processer name.}
}
\description{
Add meta data of model to mod.yaml. If there is meta data of input model, function does nothing.
}
