% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{meta_mod_output_add}
\alias{meta_mod_output_add}
\title{Add meta data of model's output.}
\usage{
meta_mod_output_add(model_name, output_file_name)
}
\arguments{
\item{model_name}{Target model name.}

\item{output_file_name}{Target output name.}
}
\description{
Add meta data of model's output to mod.yaml. If there is meta data of input model, or output, function does nothing.
}
\examples{
meta_mod_output_add('mod_sample','output_file_name.csv')
}
