% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{meta_mod_output_read}
\alias{meta_mod_output_read}
\title{Read meta data of model's output.}
\usage{
meta_mod_output_read(model_name)
}
\arguments{
\item{model_name}{Target model name.}
}
\description{
Read meta data of model's output to mod.yaml. If there is meta data of input model name or output, return NULL
}
\examples{
meta_mod_output_read('mod_sample')
}
