% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{meta_mod_read}
\alias{meta_mod_read}
\title{Read meta data of model}
\usage{
meta_mod_read(model_name = NULL)
}
\arguments{
\item{model_name}{Target model name.}

\item{processer_name}{Target processer name.}
}
\description{
Read meta data of model to mod.yaml. If model_name is NULL (default is NULL), return all model meta data as list.
}
