% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processer.R
\name{meta_proc_add}
\alias{meta_proc_add}
\title{Add proc meta data.}
\usage{
meta_proc_add(processer_name)
}
\arguments{
\item{processer_name}{Target processer name.}
}
\description{
Add proc meta data to proc.yaml   . If success to add, return TRUE, else FALSE.
}
