% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processer.R
\name{meta_proc_read}
\alias{meta_proc_read}
\title{Read proc meta data.}
\usage{
meta_proc_read(processer_name = NULL)
}
\arguments{
\item{processer_name}{Target processer name.}
}
\description{
Read proc meta data. When processer_name is null ,return all proc meta data. With processer name return list of proc meta data,you can use {name,description,input}
}
