% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_create}
\alias{model_create}
\title{Create model}
\usage{
model_create(model_name, processer_name)
}
\arguments{
\item{model_name}{Target model name.}

\item{processer_name}{Target processer name.}
}
\description{
Create model (add R srcipt & model info to mod.yaml)
}
