% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-class.R
\name{Node}
\alias{Node}
\title{Node class}
\usage{
Node(id = NA_character_, type = NA_character_, name = NA_character_,
  parameters = list(), status = "queued", output = NULL)
}
\arguments{
\item{id}{Unique node identifier}

\item{type}{Type of node (used to define internal parameters)}

\item{name}{Name of node}

\item{parameters}{List of parameters and their arguments (see details)}

\item{output}{Output of calculation for display or input into subsequent node}
}
\description{
Simple S3 class to hold node information
}
\details{
This is a simple S3 class to hold parameters relating to a single node.
The parameters argument contains information required to run the function associated with
the node.  It specifies the name and type of variable along with data required to generate
a user interface using Shiny.  For example, a numeric input may be specified as
list(name = 'n', type = 'numeric', min = 1, max = 100, step = 1, value = 50)
}

