% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineHelpers.R
\name{nodeParameters}
\alias{nodeParameters}
\title{Process node details}
\usage{
nodeParameters()
}
\value{
list of parameters associated with each node
}
\description{
Process the comments atttached to the nodes that have been added through includePackages
}

