% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{openInputs}
\alias{openInputs}
\title{Check if pipeline contains any open inputs}
\usage{
openInputs(jnt = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{jnt}{reference to htmlwidget}

\item{session}{Shiny session}
}
\description{
Determine if pipeline contains input ports with no data
}

