% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_read}
\alias{output_read}
\title{Controll output data ()}
\usage{
output_read(model_name)
}
\arguments{
\item{model_name}{Target model name.}
}
\description{
Controll output data. Read output data by model name.
}
