% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_write}
\alias{output_write}
\title{Controll output data (Save)}
\usage{
output_write(model_name)
}
\arguments{
\item{model_name}{Target model name.}
}
\description{
Controll output data. Save & Load output data by model name.
}
