% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{pipelineDFS}
\alias{pipelineDFS}
\title{Perform a depth-first search}
\usage{
pipelineDFS(jnt = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{jnt}{reference to htmlwidget}

\item{session}{Shiny session}
}
\description{
Perform a depth-first search on a pipeline.  Initial node is determined in jointPipeline.js
by identifying the first node without an input.  If all nodes have input then the first
node added is used.
}

