% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeliner.R
\docType{package}
\name{pipeliner}
\alias{pipeliner}
\alias{pipeliner-package}
\title{pipeliner: machine learning pipelines for R}
\description{
Allows you to define, fit and predict machine learning pipelines.
}

