% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_pipeline_builder.R
\name{predict.ml_pipeline}
\alias{predict.ml_pipeline}
\title{Predict method for ML pipelines}
\usage{
\method{predict}{ml_pipeline}(pipeline_object, new_data, ...)
}
\arguments{
\item{pipeline_object}{An estimated pipleine object of class \code{ml_pipeline}.}

\item{new_data}{A data.frame in which to look for variables with which to predict.}

\item{...}{Any additional arguements than need to be passed to the underlying model's predict
methods.}
}
\value{
A data.frame containing the predicted values, input variables, as well as any interim
tranformed variables.
}
\description{
Predict method for ML pipelines
}
\examples{
lm_pipeline <- ml_pipline_builder()
lm_pipeline$estimate_model(function(df) lm(y ~ x, df))
data <- data.frame(x = c(-1.26, 1.24, 0.54), y = 0.5 * c(-1.26, 1.24, 0.54) + rnorm(3))
lm_pipeline$fit(data)
predict(lm_pipeline, data)
# x          y       pred
# 1 -1.26 -0.6440193 -0.9355810
# 2  1.24  1.7833237  1.0335936
# 3  0.54 -0.5590672  0.4822247
}

