% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processer.R
\name{processer_create}
\alias{processer_create}
\title{Create processer}
\usage{
processer_create(processer_name)
}
\arguments{
\item{processer_name}{Target processer name.}
}
\description{
Create processer (Add processer script & model info to proc.yaml)
}
\examples{
processer_crate('proc_sample')
}
