% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_create}
\alias{project_create}
\title{Create project}
\usage{
project_create(project_name)
}
\arguments{
\item{project_name}{Target project name.}
}
\description{
Create function (create directory & add files.)
}
\examples{
project_create('proj_sample')
}
