% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{save_input}
\alias{save_input}
\title{Controll input data (Save & load)}
\usage{
save_input(processer_name)
}
\arguments{
\item{processer_name}{Target processer name.}
}
\description{
Controll input data. Save & Load input data by processer name
}
