% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{save_output}
\alias{save_output}
\title{Controll output data (Save & load)}
\usage{
save_output(model_name)
}
\arguments{
\item{model_name}{Target processer name.}
}
\description{
Controll output data. Save & Load output data by model name.
}
