% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch_maxn.R
\name{split_batch_maxn}
\alias{split_batch_maxn}
\title{Split vector into batches with maximum size}
\usage{
split_batch_maxn(.vec, max_content)
}
\arguments{
\item{.vec}{A vector}

\item{max_content}{The maximum lenght of each batch}
}
\value{
A list of vectors
}
\description{
Split vector into batches with maximum size
}
\examples{
\dontrun{
a <- 1:467
split_batch(a, max_content=40)
}
}
