% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBatch.R
\name{splitBatch}
\alias{splitBatch}
\title{Split a data.table into equal-sized chunks}
\usage{
splitBatch(dt)
}
\arguments{
\item{dt}{A \code{data.table} to be split.}

\item{n}{An integer indicating into how many chunks the \code{data.table} should be split.}
}
\value{
A list of \code{data.table} objects. Each element of the list is a subset
of the original \code{data.table}, with approximately \code{nrow(dt)/n} rows.
}
\description{
This function splits a \code{data.table} into approximately equal-sized parts,
based on a specified number of desired chunks. It adds a temporary grouping
variable to assign rows to chunks and returns a list of \code{data.table} objects
(without the grouping variable).
}
\examples{
library(data.table)
dt <- data.table(id = 1:465, value = rnorm(465))
chunks <- split_dt(dt, 5)

# Apply a function to each chunk with purrr
# purrr::map(chunks, some_function)
}
