% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_proc_get}
\alias{template_proc_get}
\title{Controll template}
\usage{
template_proc_get(processer_name)
}
\arguments{
\item{processer_name}{Target processer name.}

\item{model_name}{Target model name.}
}
\description{
Controlle template of {processer/model}
}
