% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_summary_table.R
\name{build_summary_table}
\alias{build_summary_table}
\title{Create table with pipeline_logs structure}
\usage{
build_summary_table()
}
\value{
A tibble with columns user_login, batch_id, symbol, status, n_rows, message
}
\description{
Create table with pipeline_logs structure
}
\examples{
\dontrun{
empty_logs <- build_summary_table()
}
}
