% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_summary_table.R
\name{build_summary_table}
\alias{build_summary_table}
\title{Create an empty summary log table with full structure}
\usage{
build_summary_table(
  user_login = Sys.getenv("USER"),
  batch_id = as.integer(Sys.time())
)
}
\arguments{
\item{user_login}{Who is running the pipeline}

\item{batch_id}{Batch identifier (e.g. based on date or increment)}
}
\value{
A tibble with the structure of pipeline_logs
}
\description{
Create an empty summary log table with full structure
}
