% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_summary_table.R
\name{build_summary_table}
\alias{build_summary_table}
\title{build_summary_table}
\usage{
build_summary_table(user_login, batch_id, n_rows, symbols)
}
\arguments{
\item{user_login}{Name of the user}

\item{batch_id}{Batch Id}

\item{n_rows}{New Rows}

\item{symbols}{list of symbols}
}
\value{
summary table
}
\description{
build_summary_table
}
