% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_predict_method_throw_error}
\alias{check_predict_method_throw_error}
\title{Validate estimate_model method returns an object with a \code{predict} method defined}
\usage{
check_predict_method_throw_error(func_return_object)
}
\arguments{
\item{func_return_object}{The object returned from the \code{estimate_model} method.}
}
\description{
Helper function that checks if the object returned from the \code{estimate_model} method has
a \code{predict} method defined for it.
}
\examples{
\dontrun{
estimation_method <- function(df) lm(eruptions ~ 0 + waiting, df)
data <- faithful
model_estimate <- estimation_method(data)
check_predict_method_throw_error(model_estimate)
# NULL
}
}

