% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the PostgreSQL Database}
\usage{
connect_db()
}
\value{
A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object to the PostgreSQL database.
}
\description{
This function wraps the internal \code{luxJob::connect_db()} function and returns a DBI connection object.
}
\examples{
con <- connect_db()
DBI::dbListTables(con)
}
