% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPipeline.R
\name{createNode}
\alias{createNode}
\title{Add a single node to stencil canvas}
\usage{
createNode(x = 0, y = 0, name = NULL, ports = c(`in` = 1, out = 1),
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{x}{position of node in canvas (x-coordinate) defined in px}

\item{y}{position of node in canvas (y-coordinate) defined in px}

\item{name}{node name (displayed as a label)}

\item{ports}{vector defining number of inputs (in) and outputs (out) for the node}

\item{session}{Shiny session}
}
\description{
Add one node to the stencil canvas
}

