% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Fetch symbols
Fetches all company symbols from the db.}
\usage{
fetch_symbols(con)
}
\arguments{
\item{con}{}
}
\description{
Fetch symbols
Fetches all company symbols from the db.
}
