% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Fetch tickers}
\usage{
fetch_symbols()
}
\value{
A \code{data.frame} of one column with all symbols
}
\description{
Fetches the tickers from the main SP500 table.
}
