% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{#' Fetch S&P500 tickers from the sp500.info table
#'
#' @param con A PostgreSQL DB connection
#'
#' @return A tibble with columns symbol and index_ts
#' @export
fetch_symbols <- function(con) {
query <- "SELECT symbol, index_ts FROM sp500.info"
symbols <- DBI::dbGetQuery(con, query)
tibble::as_tibble(symbols)
}
Fetch demo tickers (testing version)}
\usage{
fetch_symbols(con)
}
\arguments{
\item{con}{A PostgreSQL connection}
}
\value{
A tibble with real Yahoo-compatible symbols
}
\description{
#' Fetch S&P500 tickers from the sp500.info table
#'
#' @param con A PostgreSQL DB connection
#'
#' @return A tibble with columns symbol and index_ts
#' @export
fetch_symbols <- function(con) {
query <- "SELECT symbol, index_ts FROM sp500.info"
symbols <- DBI::dbGetQuery(con, query)
tibble::as_tibble(symbols)
}
Fetch demo tickers (testing version)
}
