% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Fetch demo tickers (testing version)}
\usage{
fetch_symbols(con)
}
\arguments{
\item{con}{A PostgreSQL connection}
}
\value{
A tibble with real Yahoo-compatible symbols
}
\description{
Fetch demo tickers (testing version)
}
\examples{
\dontrun{
con <- connect_db()
fetch_symbols(con)
}
}
