% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Fetch S&P500 tickers from the sp500.info table}
\usage{
fetch_symbols(con)
}
\arguments{
\item{con}{A PostgreSQL DB connection}
}
\value{
A tibble with columns symbol and index_ts
}
\description{
Fetch S&P500 tickers from the sp500.info table
}
