% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Returns all sp500 symbols and their related data from the database.
The parameter is a PqConnection object that represents an active PostgreSQL database connection.}
\usage{
fetch_symbols(con)
}
\arguments{
\item{con}{A PqConnection object that represents an active PostgreSQL connection.}
}
\value{
A data frame with the sp500 symbols and their related data.
}
\description{
Returns all sp500 symbols and their related data from the database.
The parameter is a PqConnection object that represents an active PostgreSQL database connection.
}
\examples{
\dontrun{
fetch_symbols(con)
}
}
