% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Fetch symbols}
\usage{
fetch_symbols(con)
}
\arguments{
\item{con}{Connection}
}
\value{
List of symbols (class \code{"String"})
}
\description{
Fetch symbols
}
