% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_symbols.R
\name{fetch_symbols}
\alias{fetch_symbols}
\title{Fetch distinct symbols from the database}
\usage{
fetch_symbols(con)
}
\arguments{
\item{con}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object to the PostgreSQL database.}
}
\value{
A character vector of unique symbols from the \code{sp500.info} table.
}
\description{
Connects to the database and retrieves the distinct values of \code{symbol} from the \code{sp500.info} table.
}
\examples{
con <- connect_db()
symbols <- fetch_symbols(con)
head(symbols)
}
