% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format Data
Converts the new data into the same format as the target db.}
\usage{
format_data(db, con)
}
\arguments{
\item{con}{}
}
\value{
df(newDB)
}
\description{
Format Data
Converts the new data into the same format as the target db.
}
