% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Reshape Yahoo data into long format (for database insertion).}
\usage{
format_data(data)
}
\arguments{
\item{data}{Tibble issu de tq_get(x, get = "stock.prices").}
}
\value{
Tibble en format long, trié par symbol et date.
}
\description{
Transforme un tibble tidyquant (stock.prices) en format long :
colonnes symbol, date, open, high, low, close, volume, adjusted.
}
\examples{
library(tidyquant)
data <- tq_get("AAPL", get = "stock.prices")
long_data <- format_data(data)
head(long_data)

}
