% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_data.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Insert new data into DB
Removes duplicates from df before push and then pushes it to the DB.}
\usage{
insert_new_data(newDB, con)
}
\arguments{
\item{con}{}
}
\value{
df(finalDB)
}
\description{
Insert new data into DB
Removes duplicates from df before push and then pushes it to the DB.
}
