% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_data.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Insert new data into the PostgreSQL table}
\usage{
insert_new_data(.tidy_data)
}
\arguments{
\item{.tidy_data}{A tidy tibble containing stock market data}
}
\value{
A message confirming the insertion or not of the records
}
\description{
Insert new data into the PostgreSQL table
}
