% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_data.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Insert new metrics into student_ibtissam.data_sp500}
\usage{
insert_new_data(con, data, schema = "student_ibtissam")
}
\arguments{
\item{con}{A PostgreSQL connection}

\item{data}{A tibble with index_ts, date, metric, value}

\item{schema}{PostgreSQL schema name (default = student_ibtissam)}
}
\value{
TRUE if insertion completed
}
\description{
Insert new metrics into student_ibtissam.data_sp500
}
