% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_data.R
\name{insert_new_data}
\alias{insert_new_data}
\title{insert_new_data}
\usage{
insert_new_data(conn, index_ts, data)
}
\arguments{
\item{conn}{Connection}

\item{index_ts}{index_ts}

\item{data}{Data}
}
\description{
insert_new_data
}
