% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_data.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Insert New Observations into PostgreSQL Table}
\usage{
insert_new_data(con, formattedTable = list())
}
\arguments{
\item{con}{A PostgreSQL DBI connection object.}

\item{formattedTable}{A list of \code{tibble}s containing cleaned and reshaped data.}
}
\value{
A list of \code{tibble}s containing only the new observations inserted into the database.
}
\description{
This function identifies and inserts new data points from a formatted list of \code{tibble}s into a
PostgreSQL table. It compares with existing records based on \code{index_ts}, \code{date}, and \code{metric}.
}
\examples{
\dontrun{
con <- connect_db()
new_data <- insert_new_data(con, formattedTable)
}
}
