% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summary.R
\name{log_summary}
\alias{log_summary}
\title{Log summary}
\usage{
log_summary(
  .user_login = "EriF",
  .start = lubridate::today() - 6,
  .end = lubridate::today() - 1,
  n_stocks_per_batch = 31,
  .symbols,
  .message
)
}
\arguments{
\item{.user_login}{Optional. A String containing the username}

\item{.start}{Optional. A Date with the starting date}

\item{.end}{Optional. A Date for the last day}

\item{n_stocks_per_batch}{Optional. An Integer with the number of ticker per batch}

\item{.symbols}{A Vector with all the tickers}

\item{.message}{A String message that has been prompted by the insert_new_data}
}
\value{
A tibble with the log
}
\description{
Log summary
}
\examples{
\dontrun{
pipeline_logs <- log_summary(.symbols = symbols_vector, .message = message)
}
}
