% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summary.R
\name{log_summary}
\alias{log_summary}
\title{log_summary}
\usage{
log_summary(summary_tbl, n_rows, status, symbols)
}
\arguments{
\item{summary_tbl}{Summary table}

\item{n_rows}{New rows}

\item{status}{Status}

\item{symbols}{List of symbols}
}
\description{
log_summary
}
