% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_write}
\alias{output_write}
\title{Controll output data (Write)}
\usage{
output_write(x, model_name)
}
\arguments{
\item{x}{Target output data.}

\item{model_name}{Target model name.}

\item{write_func}{How to save x. e.g) write.csv()}
}
\description{
Controll output data. Write output data by output_data name & model name
}
