% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_create.R, R/modeling.R, R/preprocessing.R
\docType{package}
\name{pipelineR-package}
\alias{pipelineR}
\alias{pipelineR-package}
\title{create folders for data analytics}
\arguments{
\item{Data}{Data to use the model}

\item{target}{dependent variable.}

\item{model_type}{which model to use: "lm" or "logistic" or "poisson"}

\item{id_name}{if Data has ID columns, select it...、}

\item{use_features}{}
}
\value{
model
}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\examples{
dir_create()
}
\keyword{internal}
