% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_create.R
\docType{package}
\name{pipelineR-package}
\alias{hello-package}
\title{create folders for data analytics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
