% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_create.R, R/modeling.R, R/preprocessing.R
\docType{package}
\name{pipelineR-package}
\alias{pipelineR}
\alias{pipelineR-package}
\title{create folders for data analytics}
\arguments{
\item{R_proj}{make .Rproj file. default is TRUE}

\item{dir}{if define working directory, set as character. default is NULL}

\item{Data}{Data to use the model}

\item{target}{dependent variable.}

\item{model_type}{which model to use: "lm" or "logistic" or "poisson"}

\item{id_name}{if Data has ID columns, select it...、}

\item{use_features}{}
}
\value{
model
}
\description{
The pipeline of basic statistical modeling and output

The pipeline of basic statistical modeling and output

The pipeline of basic statistical modeling and output
}
\examples{
dir_create()
}
\keyword{internal}
