% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineR-package.R
\docType{package}
\name{pipelineR-package}
\alias{pipelineR}
\alias{pipelineR-package}
\title{pipelineR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yusuke Tanaka \email{yusuke.tanaka.keio@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
