% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_summary_table.R
\name{push_summary_table}
\alias{push_summary_table}
\title{Push summary table to the DB}
\usage{
push_summary_table(.summary_table, con)
}
\arguments{
\item{.summary_table}{A tibble that will be uploaded to the database}

\item{con}{A connection}
}
\value{
A message
}
\description{
Push summary table to the DB
}
\examples{
\dontrun{
push_summary_table(.summary_table = pipelog_table)
}
}
