% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_summary_table.R
\name{push_summary_table}
\alias{push_summary_table}
\title{Push log entries into pipeline_logs}
\usage{
push_summary_table(con, log_tbl, schema = "student_ibtissam")
}
\arguments{
\item{con}{DB connection}

\item{log_tbl}{Tibble with logs}

\item{schema}{Schema name}
}
\value{
TRUE
}
\description{
Push log entries into pipeline_logs
}
