% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_summary_table.R
\name{push_summary_table}
\alias{push_summary_table}
\title{Push logs to student_ibtissam.pipeline_logs}
\usage{
push_summary_table(con, log_tbl, schema = "student_ibtissam")
}
\arguments{
\item{con}{DB connection}

\item{log_tbl}{Log tibble}

\item{schema}{Schema name}
}
\value{
TRUE if successful
}
\description{
Push logs to student_ibtissam.pipeline_logs
}
