% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_summary_table.R
\name{push_summary_table}
\alias{push_summary_table}
\title{push_summary_table}
\usage{
push_summary_table(conn, summary_tbl)
}
\arguments{
\item{conn}{Connection}

\item{summary_tbl}{Summary Table}
}
\value{
void
}
\description{
push_summary_table
}
