% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_summary_table.R
\name{push_summary_table}
\alias{push_summary_table}
\title{Push Summary Table to PostgreSQL}
\usage{
push_summary_table(summaryTable = tibble(), con = con)
}
\arguments{
\item{summaryTable}{A \code{tibble} containing the summary of new observations that were processed.
The table must have the same structure as the \code{pipeline_logs} table in the PostgreSQL database.}

\item{con}{A connection to the database.}
}
\value{
NULL This function does not return anything. It performs an append operation to the database.
}
\description{
This function appends the summary table (containing batch processing details) to a PostgreSQL table.
The table is stored under a schema named with the current user's environment variable \code{PG_USER}.
}
\examples{
\dontrun{
push_summary_table(summaryTable)
}
}
