% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{Split df
Splits the df into batches of 25 items.}
\usage{
split_batch(symbols)
}
\arguments{
\item{symbols}{}
}
\value{
list(df("batches"), integer("nrows"))
}
\description{
Split df
Splits the df into batches of 25 items.
}
