% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{Split vector into determined number of batches}
\usage{
split_batch(.vec, n_groups)
}
\arguments{
\item{.vec}{A vector}

\item{n_groups}{The number of desired batches}
}
\value{
A list of vectors
}
\description{
Returns a list containing multiple vectors (n_groups). It splits the lenght
of the vectors as fair as possible.
}
\examples{
\dontrun{
a <- 1:467
split_batch(a, n_groups=15)
}
}
