% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{Split Items into Batches}
\usage{
split_batch(items, batch_size)
}
\arguments{
\item{items}{A vector containing items to split.}

\item{batch_size}{Integer. Max number of items per batch.}
}
\value{
A list, where each element is a batch. Returns empty list if items is empty.
}
\description{
Divides a vector into a list of smaller vectors (batches).
}
