% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{Split tickers into batches}
\usage{
split_batch(symbols, batch_size = 20)
}
\arguments{
\item{symbols}{A tibble of tickers (from fetch_symbols)}

\item{batch_size}{Number of tickers per batch}
}
\value{
A list of tibbles
}
\description{
Split tickers into batches
}
