% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{Returns the complete stock symbol table horizontally split up into}
\usage{
split_batch(tbl, batch_size = 25)
}
\arguments{
\item{tbl}{A data frame that represents the complete data table of the stock symbols.}

\item{batch_size}{A numeric parameter that represents the batch size.}
}
\value{
A list of data frames of the split-up symbol table.
}
\description{
Returns the complete stock symbol table horizontally split up into
}
\examples{
\dontrun{
split_batch(df, 50)
split_batch(df)
}
}
