% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{split_batch}
\usage{
split_batch(data, batch_size)
}
\arguments{
\item{data}{Data}

\item{batch_size}{Size of batches}
}
\value{
batches
}
\description{
split_batch
}
