% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_batch.R
\name{split_batch}
\alias{split_batch}
\title{Split a Vector into Random Batches}
\usage{
split_batch(vec = symbols, batch_size = 10)
}
\arguments{
\item{vec}{A vector to be split into batches (e.g., character, numeric).}

\item{batch_size}{An integer indicating the size of each batch.}
}
\value{
A named list of vectors, each representing a batch.
}
\description{
Randomly splits a vector into a list of smaller batches of a specified size.
}
\examples{
symbols <- c("AAPL", "GOOG", "MSFT", "AMZN", "TSLA")
split_batch(symbols, batch_size = 2)
}
