% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_pipeline.R
\name{start_pipeline}
\alias{start_pipeline}
\title{Orchestrator
Orchestrates all of the other functions.}
\usage{
start_pipeline()
}
\value{
nothing.
}
\description{
Orchestrator
Orchestrates all of the other functions.
}
