% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_pipeline.R
\name{start_pipeline}
\alias{start_pipeline}
\title{Start pipeline}
\usage{
start_pipeline(.from = Sys.Date() - 7, .to = Sys.Date(), .batch_size = 31)
}
\arguments{
\item{.from}{Optional. Start date}

\item{.to}{Optional. End date}

\item{.batch_size}{Optional. Number of tickers per batch}
}
\value{
Undefined
}
\description{
Start pipeline
}
\examples{
\dontrun{
start_pipeline()
}
}
