% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_pipeline.R
\name{start_pipeline}
\alias{start_pipeline}
\title{Run the full data pipeline}
\usage{
start_pipeline(from, to, batch_size = 10, user_login = Sys.getenv("USER"))
}
\arguments{
\item{from}{Start date (Yahoo Finance)}

\item{to}{End date (Yahoo Finance)}

\item{batch_size}{Number of tickers per batch}

\item{user_login}{Name of the user running the pipeline}
}
\description{
Run the full data pipeline
}
