% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_pipeline.R
\name{start_pipeline}
\alias{start_pipeline}
\title{start_pipeline}
\usage{
start_pipeline(user, from, to, batch_size)
}
\arguments{
\item{user}{user}

\item{from}{from}

\item{to}{to}

\item{batch_size}{batch_size}
}
\value{
void
}
\description{
start_pipeline
}
