% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_query_data.R
\name{yahoo_query_data}
\alias{yahoo_query_data}
\title{Yahooquery
Queries the yahoo db for stock prices up until yesterday
and then extracts the latest stock price for each item.}
\usage{
yahoo_query_data(batch)
}
\arguments{
\item{batch}{}
}
\value{
df("latest stock prices")
}
\description{
Yahooquery
Queries the yahoo db for stock prices up until yesterday
and then extracts the latest stock price for each item.
}
