% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_query_data.R
\name{yahoo_query_data}
\alias{yahoo_query_data}
\title{Query data from Yahoo finance}
\usage{
yahoo_query_data(
  .vec,
  n_stocks_per_batch = 50,
  .start,
  .end = lubridate::today() - 1
)
}
\arguments{
\item{.vec}{A vector with the tickers}

\item{n_stocks_per_batch}{An integer that sets the maximum number of tickers per batch}

\item{.start}{A date "YYYY-MM-DD" for the beginning of the data extraction}

\item{.end}{A date "YYYY-MM-DD" for the end of the data extraction}
}
\description{
Query data from Yahoo finance
}
\examples{
\dontrun{
tickers <- c("A", "AAPL", "ABBV")
yahoo_query_data(tickers, n_stocks_per_batch = 1, .start = "2025-01-15")
}
}
