% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_quety_data.R
\name{yahoo_query_data}
\alias{yahoo_query_data}
\title{Query Yahoo Finance for OHLCV data}
\usage{
yahoo_query_data(batch, from, to)
}
\arguments{
\item{batch}{A tibble of tickers (from split_batch)}

\item{from}{Start date (YYYY-MM-DD)}

\item{to}{End date (YYYY-MM-DD)}
}
\value{
A named list of tibbles, one per ticker
}
\description{
Query Yahoo Finance for OHLCV data
}
