% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_query_data.R
\name{yahoo_query_data}
\alias{yahoo_query_data}
\title{Returns a data frame that contains stock data for the requested tickers in the given date window.}
\usage{
yahoo_query_data(symbols, from, to)
}
\arguments{
\item{symbols}{A list of stock symbols.}

\item{from}{A date value as a start date of the requested period.}

\item{to}{A date value as an end date of the requested period.}
}
\value{
A data frame with the stock data for the requested tickers in the given date window.
}
\description{
Returns a data frame that contains stock data for the requested tickers in the given date window.
}
\examples{
\dontrun{
yahoo_query_data(c("AAPL", "A", "CNP", "CRL"), "2015-01-02", "2015-01-04")
}
}
