% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_query_data.R
\name{yahoo_query_data}
\alias{yahoo_query_data}
\title{yahoo_query_data}
\usage{
yahoo_query_data(ticker, start_date, end_date)
}
\arguments{
\item{ticker}{Ticker}

\item{start_date}{Start date}

\item{end_date}{End date}
}
\value{
data
}
\description{
yahoo_query_data
}
