% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_query_data.R
\name{yahoo_query_data}
\alias{yahoo_query_data}
\title{Query Yahoo Finance Data for a List of Symbol Batches}
\usage{
yahoo_query_data(symbolsBatches, from = lubridate::now() - lubridate::days(5))
}
\arguments{
\item{symbolsBatches}{A list of character vectors, each representing a batch of stock symbols.}

\item{from}{A \code{Date} or \code{POSIXct} object specifying the start date for data retrieval. Defaults to 5 days ago.}
}
\value{
A list of \code{tibble}s with stock price data for each batch.
}
\description{
Retrieves stock price data for each batch of symbols using \code{tq_get()} from the \code{tidyquant} package.
}
\examples{
symbols <- c("AAPL", "GOOG", "MSFT")
batches <- split_batch(symbols, batch_size = 2)
data <- yahoo_query_data(batches)
}
